test: refactoring tests into separate test suites
