+++ test/querySelector.spec.js
@@ -6,7 +6,8 @@
  */
 /* global it describe */
 
-import { querySelectorAll, querySelector, getElementsByClassName } from '../src/index.js'
+import { querySelectorAll, querySelector } from '../src/index.js'
+import { assertSelectorEqual, withDom, simplifyElement, simplifyElements } from './utils.js';
 import assert from 'assert'
 import simpleLight1 from './fixtures/simple1/light.html'
 import simpleShadow1 from './fixtures/simple1/shadow.html'
@@ -51,72 +52,26 @@ import unusualSelectorsShadow1 from './fixtures/unusualSelectors1/shadow.html'
 import classNamesLight1 from './fixtures/classNames1/light.html'
 import classNamesShadow1 from './fixtures/classNames1/shadow.html'
 
-function withDom (html, cb) {
-  const iframe = document.createElement('iframe')
-  document.body.appendChild(iframe)
-  const iframeDocument = iframe.contentWindow.document
-  iframeDocument.open('text/html', 'replace')
-  iframeDocument.write(html)
-  iframeDocument.close()
-  try {
-    cb(iframeDocument)
-  } finally {
-    document.body.removeChild(iframe)
-  }
-}
-
-function simplifyElement (element) {
-  if (!element) {
-    return undefined
-  }
-  return {
-    tagName: element.tagName,
-    classList: [...element.classList]
-  }
-}
-
-function simplifyElements (elements) {
-  return [...elements].map(simplifyElement)
-}
-
-function stringify (obj) {
-  if (typeof obj === 'undefined') {
-    return obj
-  }
-  return JSON.stringify(obj)
-}
-
-function assertResultEqual (selector, actual, expected, returnsCollection) {
-  if (returnsCollection) {
-    actual = simplifyElements(actual)
-  } else {
-    actual = simplifyElement(actual)
-    expected = expected[0]
-  }
-  assert.deepStrictEqual(actual, expected,
-    `Selector failed: ${stringify(selector)}, ${stringify(actual)} !== ${stringify(expected)}`)
-}
-
 function testSelectors (lightDom, shadowDom, tests) {
   tests.forEach(({ selector, expected }) => {
     it('light DOM - qSA', () => {
       withDom(lightDom, context => {
-        assertResultEqual(selector, context.querySelectorAll(selector), expected, true)
+        assertSelectorEqual(selector, context.querySelectorAll(selector), expected, true)
       })
     })
     it('shadow DOM - qSA', () => {
       withDom(shadowDom, context => {
-        assertResultEqual(selector, querySelectorAll(selector, context), expected, true)
+        assertSelectorEqual(selector, querySelectorAll(selector, context), expected, true)
       })
     })
     it('light DOM - qS', () => {
       withDom(lightDom, context => {
-        assertResultEqual(selector, context.querySelector(selector), expected, false)
+        assertSelectorEqual(selector, context.querySelector(selector), expected, false)
       })
     })
     it('shadow DOM - qSA', () => {
       withDom(shadowDom, context => {
-        assertResultEqual(selector, querySelector(selector, context), expected, false)
+        assertSelectorEqual(selector, querySelector(selector, context), expected, false)
       })
     })
   })
@@ -952,28 +907,28 @@ describe('basic test suite', function () {
     it('light DOM - getElementsByClassName', () => {
       const classNames = 'container main'
       withDom(classNamesLight1, context => {
-        assertResultEqual(classNames, context.getElementsByClassName(classNames), expected, true)
+        assertSelectorEqual(classNames, context.getElementsByClassName(classNames), expected, true)
       })
     })
 
     it('shadow DOM - getElementsByClassName', () => {
       const classNames = 'container main'
       withDom(classNamesShadow1, context => {
-        assertResultEqual('.container.main', getElementsByClassName(classNames, context), expected, true)
+        assertSelectorEqual('.container.main', getElementsByClassName(classNames, context), expected, true)
       })
     })
 
     it('light DOM - getElementsByClassName with multiple spaces', () => {
       withDom(classNamesLight1, context => {
         const classNames = 'container      main'
-        assertResultEqual(classNames, context.getElementsByClassName(classNames), expected, true)
+        assertSelectorEqual(classNames, context.getElementsByClassName(classNames), expected, true)
       })
     })
 
     it('shadow DOM - getElementsByClassName with multiple spaces', () => {
       withDom(classNamesShadow1, context => {
         const classNames = 'container      main'
-        assertResultEqual(classNames, getElementsByClassName(classNames, context), expected, true)
+        assertSelectorEqual(classNames, getElementsByClassName(classNames, context), expected, true)
       })
     })
 
@@ -984,7 +939,7 @@ describe('basic test suite', function () {
         
         
         container`
-        assertResultEqual(classNames, context.getElementsByClassName(classNames), expected, true)
+        assertSelectorEqual(classNames, context.getElementsByClassName(classNames), expected, true)
       })
     })
 
@@ -995,7 +950,7 @@ describe('basic test suite', function () {
         
         
         container`
-        assertResultEqual(classNames, getElementsByClassName(classNames, context), expected, true)
+        assertSelectorEqual(classNames, getElementsByClassName(classNames, context), expected, true)
       })
     })
 
@@ -1004,14 +959,14 @@ describe('basic test suite', function () {
     it('light DOM - getElementsByClassName with tabs', () => {
       withDom(classNamesLight1, context => {
         const classNames = 'main	container'
-        assertResultEqual(classNames, context.getElementsByClassName(classNames), expected, true)
+        assertSelectorEqual(classNames, context.getElementsByClassName(classNames), expected, true)
       })
     })
 
     it('shadow DOM - getElementsByClassName with tabs', () => {
       withDom(classNamesShadow1, context => {
         const classNames = 'main	container'
-        assertResultEqual(classNames, getElementsByClassName(classNames, context), expected, true)
+        assertSelectorEqual(classNames, getElementsByClassName(classNames, context), expected, true)
       })
     })
 
@@ -1021,7 +976,7 @@ describe('basic test suite', function () {
         
             	
         `
-        assertResultEqual(classNames, context.getElementsByClassName(classNames), expected, true)
+        assertSelectorEqual(classNames, context.getElementsByClassName(classNames), expected, true)
       })
     })
 
@@ -1031,7 +986,7 @@ describe('basic test suite', function () {
         
             	
         `
-        assertResultEqual(classNames, getElementsByClassName(classNames, context), expected, true)
+        assertSelectorEqual(classNames, getElementsByClassName(classNames, context), expected, true)
       })
     })
